Make all methods take widgets by & instead of &mut
All of the widget types are wrappers around recounted shared references
and contain no data we would ever mutate. There's no meaningful difference
between a shared (&) and a mutable (&mut) reference to a widget, the latter
doesn't guarantee the uniqueness of the reference it wraps.

We have a practical reason to change this now: a `Fn` closure can't call any
mutating methods on the variables moved into it so the status quo would
artificially limit the usefulness of signal handlers proposed in #33.