chore: Update clippy settings somewhat
Clippy attributes are now namespaced and should be used as such,
so transform the old cfg_attr stuff into clippy:: lint names.

Also adds a few more to the auto list and moves one to the global
list to quieten warnings which we can't do much about as it's the
way GTK mostly works.

Signed-off-by: Daniel Silverstone <dsilvers@digital-scurf.org>